First script for database Dump.
The script is for use pg_dump in the database name list in the class
and if the connection to "MARCAJE" folder (FTP connection) is active
(if python found my .verificador file I let in the folder) then
copy the database dump at that Server.

That was the server of backups, thats why I send the backups there in the script.

v1.0.0